[refactor/#52] 드롭다운 구조 정리
